Makefile: on Linux use -fvisibility=hidden for everything, not just DLLs
On typical CPU architectures like x86 and ARM, symbols in the executable
are not exported unless you specifically build with `-rdynamic`, but on
some architectures every `extern` symbol is exported into the dynamic
symbol table by default. This means they will "interpose" in front of
symbols of the same name in a library, which is very bad if it was
unintentional (for example if the address of `vmCvar_t sv_fps` in
OpenArena's `qagame.so` is unexpectedly overwritten with the address
of `cvar_t sv_fps` in the ioq3ded executable, a global variable of a
different type with a different in-memory layout).

We can avoid this by building everything with hidden visibility, so that
only symbols that are intentionally exported (such as `vmMain` and
`dllEntry`) are present in the dynamic symbol table. This results in
approximately the same semantics that a Windows developer would expect,
and in practice is usually what is desired.

Resolves: https://github.com/ioquake/ioq3/issues/765
Signed-off-by: Simon McVittie <smcv@debian.org>

Note: This was submitted to ioquake3 but alternative solutions were used.
It was cherry-picked to Lilium Arena without the author's involvement.